chore(deps): update `Cargo.lock` dependencies
This commit includes a substantial update to the Cargo.lock, with the addition of a slew of new dependencies and checksums. The update essentially includes new and modified dependencies necessary for the proper functioning of the project.

These modifications are critical as they help ensure the project's success by incorporating the latest and most stable versions of each dependency. Moreover, these updates also help to mitigate any potential vulnerabilities found in older versions of dependencies by keeping them up-to-date.

The additions span a broad range of functionality and seem to be mostly related to web scraping and communication (ex. reqwest, ftm-http and web-sys), as well as lower level systems operations such as process control (ex. blueox, proc-macro2). This would suggest an expansion of the project's scope or functionality.