refactor: update Discord bot methods to async
Refactored the prepare(), process(), and dispatch() methods of the Discord bot from synchronous to asynchronous. This change is done to improve the performance and responsiveness of the bot. Furthermore, this change replaces the use of the request blocking get with non-blocking asynchronous get method in PropHouseDiscordBot as part of the asynchronous refactoring. Also, adjusted the main function to accommodate the async nature of the bot methods.