refactor: add Serialize trait to structs and update database operations
The Serialize trait has been added to all data structures to enable serialization of these objects to save to a database.
Previously, we were printing the Votes for each Proposal, instead, they are now stored in a sled database for persistent storage which will permit analysis and reporting in the future. This update aims to streamline data usage for improved performance.