chore(deps): add `serenity` crate to `Cargo.toml` for discord event handling
The serenity crate was added to the Cargo.toml file in order to implement a new discord bot, PropLotDiscordBot. This change was necessary as it improves the bot performance and debugging capabilities by allowing for more sophisticated control in processing and dispatching events in the main function. The effectiveness of this implementation will be evaluated through continuous monitoring and testing.