refactor: add serde attribute for camelCase naming convention
Added the `serde(rename_all = "camelCase")` attribute to `IdeaResponse`, `Idea`, `IdeaCount`, `IdeaVote` and `IdeaVoter` structs for serialization and deserialization. This change was made to comply with JSON naming conventions and to allow consistent data exchange between front-end and back-end. The attribute modifies the output of Serialize and input of Deserialize implementations. The variable names have now been changed to snake_case compliant with Rust's naming conventions, ensuring readability of struct fields. Commented out an initialisation of `_idea_lifecycle_handlers` array as it led to an unused variable warning. Removed a call to `get_all_ideas().await?;` in `prepare` function as it was not used.