refactor: improve cache setup and events handling in Discord bot
This commit includes several major improvements in the prop_lot_discord_bot module. Specifically the cache setup functions for ideas was added which provides better handling of the cache before accessing it. The function has_idea_cache_setup() is introduced to check if the cache was set up or not. Also, this commit moves from manual iteration over tasks to leveraging futures::future::join_all to improve performance and readability. Error handling was also bolstered with replacement of manual .unwrap calls with the ? operator. Redundant code was also removed and we introduced improvements in async functions leading to better performance and cleaner code. The process function was improved to offers more details about the event's progress and functionality.