chore: add GraphQL schema for Prop Lot
A new file `prop_lot_schema.graphql` has been added to implement the property lot functionality in the application. This includes types for users, votes, comments, and ideas. It also includes schema for operations like getting a user, adding a vote, submitting an idea and associated comments. This change is part of an effort to use GraphQL as the query language in order to manage the data at the backend more efficiently.