chore: add the PropLot query to GraphQL
Created a new GraphQL query for a property lot, named PropLot, in the graphql/queries directory. This query will fetch data related to ideas such as id, title, vote count, and more. The need for this query arose to support the functionality of retrieving detailed data around the idea, including creator details and voter-related information.