refactor: improve and modularize Prop Lot fetching logic
The logic for fetching PropLot ideas has been moved from the main.rs file to a new fetcher.rs file within a new prop_lot directory. This is done to encapsulate the related functionalities and improve the readability of the code. Further, the GraphQL related entities such as 'Query' and 'Idea' have been renamed. This would improve understanding of the code and enhance maintainability.