chore: add `git-cliff` configuration and initial CHANGELOG.md
A default configuration file (cliff.toml) for git-cliff has been added along with the project's initial CHANGELOG.md. Git-cliff is a tool used for generating changelogs from the Git history. The configuration file consists of a number of settings that organize and determine how the commits will be grouped and documented in the CHANGELOG.md file. This is an important step towards managing and documenting all notable changes made to the project in a readable and structured manner.