refactor: switch storage engine from `sled` to `rocksdb`
The changes include replacing sled DB with rocksdb for cache storage in file cache.rs. Rocksdb has been chosen due to its performance advantage and extended functionality. The changes also imply adjustments to certain function calls due to the differences between the two DB engines. Note that with rocksdb an option is set to create a database if not already existent.