refactor: move handler functions to DiscordHandler structs
The previous design of prop_house handler had several async functions handling auctions, votes etc individually. Functions were configured with environment variables repeatedly, which made the code harder to read and manage. In this commit, refactoring the handler code to an object-oriented structure using a new "DiscordHandler" struct has made the code more organized and readable. Each handler function is now a method of the DiscordHandler struct and the `start` method initializes an instance of this struct and reuses it, enhancing performance efficiency. Centralizing the handling of base_url, http, and webhook also minimizes the chance of errors.

fixup! refactor: move handler functions to DiscordHandler structs