chore: add `node_modules` to `.gitignore`
This change is incorporated to ensure we are not tracking or commiting the node_modules folder. This is an important change in order to reduce the clutter and size of our repository as this folder can potentially contain thousands of npm packages which are not necessary to be in our version control.