refactor: remove caching system from project
The cache.rs and associated cacher.rs files directly related to it in the 'prop_house' and 'prop_lot' directories were removed. The decision to remove the caching system from the project was made because it was found to be an unnecessary optimization. Caching is typically used for performance optimization to speed up software apps by storing data in memory that can be re-used. However, upon reviewing the needs of our application, we determined that our application's data needs were simple enough that we did not require a caching system and this would by extension reduce code complexity.