refactor: update fetcher functions to use worker env
The fetcher functions in both prop_house and prop_lot modules have been refactored to use worker::Env instead of std::env. This change simplifies the handling of environment variables and reduces dependencies. It also removes several lines of repetitive code and unifies the way environment variables are accessed across the modules, therefore improving the readability and maintainability of the code. The dependency on std::time::Duration was removed as it was no longer required.