refactor: add caching functionality to enhance performance
A new struct 'Cache' was introduced in the crate 'cache.rs' which acts as a key-value (KV) store. It abstracts the functionalities of a KV store with methods for putting and getting data. The data is serialized and deserialized using the serde library. This cache layer will boost the performance by reducing the need for direct database calls.