refactor: update DiscordHandler to use a custom in-memory cache system
In this commit, the previous Serenity-based webhook system and Cacher system have been replaced with a more general reqwest Client and custom Cache in 'src/prop_lot/handler.rs' and 'src/prop_house/handler.rs'. Now, we are using the Client and Cache system which is present in 'worker'.
This refactoring brings consistency in the usage of resources available in the worker environment, and simplifies the dependency tree by eliminating the need for modules that were only used for executing the Discord webhooks and caching.
The organization level webhook URL is fetched from the environment variables to initialize the DiscordHandler.
Error handling for unsuccessful webhook requests has been improved, and unsuccessful cache retrievals are logged as errors but do not interrupt the process.
The in-memory caching system noticeably impacts the run time performance.