refactor: update data type in prop lot module
Updated the data type from Vec<Vote> to Vec<Comment> in the cache.get method of the prop_lot module. Also added Comment from the fetcher module at the top of the module. This change is done to reflect an alteration in the data structure used for caching, keeping the cache more consistent with the type of data being processed, i.e. comment data.