chore: update `deploy` script in `package.json`
The "deploy" script in the package.json file has been modified to specify the deployment environment. Previously, the deploy script was running without any set environment. To keep consistency with other scripts and to ensure the process occurs in the right environment, the deploy script now includes "--env dev" argument.