refactor: add Handler trait to `prop_lot` module
This update introduces the 'Handler' trait to the 'prop_lot' module. It includes the implementation of three async methods - 'handle_new_idea', 'handle_new_vote', and 'handle_new_comment'. These methods aim to provide the functionality of handling new entries in idea, vote, and comment respectively. This new feature was added to improve event handling in the application.