refactor: integrate Handler trait with PropLot struct
The existing implementation of PropLot has been made more generic by allowing it to accommodate any object implementing the Handler trait, instead of being limited to the DiscordHandler. This change involves significant modifications to both 'src/lib.rs' and 'src/prop_lot/mod.rs'. The utility of this generic implementation will be realized when working with different type of handlers in future. This is a work in progress.