refactor: add proposal and vote handling in DiscordHandler of meta gov
Added the ability to handle new proposals and votes in the DiscordHandler. When a new proposal or vote is detected, appropriate information is logged, data is formatted and a webhook is used to broadcast this information. Also, imported the chrono library to handle timestamps. This will enable real-time tracking of proposals and votes within the MetaGov system through Discord.