chore: add meta gov dev environment variables
The commit introduces META_GOV_BASE_URL for the dev environment in wrangler.toml. This new variable is necessary to connect to the nounsdao voting platform and ensure its proper functionality in the development environment. The change aims to accelerate the development process and increase overall productivity.src/lib.rs' related to PropLot and PropHouse initializations have been temporarily commented out. This is likely for debugging the recent changes or dealing with dependencies.