refactor: add function to extract proposal information on meta gov handler
This commit introduces a new function `extract_proposal_info` to the `DiscordHandler` structure in the `meta_gov/handler/discord.rs` file which extracts a Proposal's ID and title by using regex. This newly extracted information is then used in any webhook execution or in handling any new proposals. This was done to improve the integrity and accuracy of the information passed around in the codebase.