ci: add pnpm package manager installation step to deploy workflow
An extra step was added to the GitHub Actions deployment workflow to install the pnpm package manager. This change was necessary to ensure that pnpm, which is required for managing the project's dependencies, is installed in the environment before the build takes place. The version that will be installed is always the latest one.