refactor: replace unwrap with error handling in farcaster handlers
Replaced instances where `unwrap` was used to directly access inner values with a more robust error handling method. This change aims to enhance the error handling mechanism in `farcaster.rs` files related to `lil_nouns`, `meta_gov` and `prop_house` handlers. Now, instead of causing a panic with a less informative message, it will return an error and a more informative "Proposal not found in the funding list"/ "Auction not found in the funding list" error message when a proposal/auction is not found.