refactor: handle default values for new and old prices in second market
In the second_market module, the handling of the default values for new and old prices have been improved. The default value for old price is now returned when the cache does not contain an old price value. Similarly, the default value for new price is returned when the price is not defined in the new floor. The condition to handle new floors has also been updated to check for "new-order" type along with the inequality of new and old prices.